+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Bugfixes
 * Fix for ``DblSphere`` halo exclusion model.
 * Fix for when no tracer profile/concentration params are given. Now use the halo params
   iff the tracer model is equal to the halo one.
+* Fix for halo profile numerical FT.
 
 v1.4.5
 ------